Refactor dtcs into a ecu->dtc_list table of tables. Move them out into separate files.
